pyproject.toml
[project]
name = "genai-assistent-schubert-ss-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.21",
    "datasets>=4.1.1",
    "docx2txt>=0.9",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.6",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.24",
    "langchainhub>=0.1.21",
    "matplotlib>=3.10.6",
    "numpy>=2.3.0",
    "openai>=1.88.0",
    "openai-agents>=0.3.1",
    "pandas>=2.3.0",
    "pinecone-client>=6.0.0",
    "pypdf>=5.6.0",
    "python-dotenv>=1.1.0",
    "ragas>=0.3.5",
    "requests>=2.32.4",
    "streamlit>=1.50.0",
    "tiktoken>=0.9.0",
    "vadersentiment>=3.3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
exclude = [
    ".git",
    ".gitignore",
    "__pycache__",
    "*.egg-info",
    ".nox",
    ".mypy_cache",
    ".ruff_cache",
    ".pytest_cache",
    ".venv",
    "venv",
    ".vscode",
    ".idea"
]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "FA",
    "PT",
    "TD",
    "BLE",
    "COM",
    "CPY",
    "DTZ",
    "ERA",
    "FIX",
    "INP",
    "PGH",
    "PLR",
    "RET",
    "T20",
    "TRY",
    "C408",
    "D104",
    "D107",
    "D203",
    "D212",
    "N806",
    "S101",
    "S311",
    "PD901",
    "ANN002",
    "ANN003",
    "ANN401",
    "RUF012",
    "RUF100",
    "ISC001",
    "E501",
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.mccabe]
max-complexity = 15


[tool.pyright]
typeCheckingMode = "standard"
pythonVersion = "3.12"
include = ["src"]
extraPaths = ["src"]
exclude = [
    "**/__pycache__",
    "**/.pytest_cache",
    "**/.ruff_cache",
    "**/.venv",
    "**/venv"
]